requirements.txt
fastapi==0.115.0
uvicorn==0.30.1
pillow==10.4.0
pytesseract==0.3.10
python-multipart==0.0.9
